requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0 